dash-dbx-writeback/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-annotate --no-hashes --format requirements-txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
dash==3.0.4
dash-ag-grid==31.3.1
dash-iconify==0.1.2
dash-mantine-components==2.0.0
databricks-sdk==0.56.0
databricks-sql-connector==4.0.3
et-xmlfile==2.0.0
flask==3.0.3
google-auth==2.40.3
greenlet==3.2.3 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
lz4==4.4.4
markupsafe==3.0.2
narwhals==1.41.1
nest-asyncio==1.6.0
numpy==2.3.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
plotly==6.1.2
psycopg==3.2.9
psycopg-binary==3.2.9 ; implementation_name != 'pypy'
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sqlalchemy==2.0.41
thrift==0.20.0
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
werkzeug==3.0.6
zipp==3.22.0
